rename types _types
